backend/backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# LangChain and AI
langchain==0.0.340
langchain-community==0.0.1
llama-cpp-python==0.2.20
sentence-transformers==2.2.2
huggingface-hub==0.19.4

# Online LLM Providers
openai==1.3.0
anthropic==0.7.0
google-generativeai==0.3.0
cohere==4.37

# OCR
pytesseract==0.3.10
Pillow==10.1.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0